Renamed Tournament  & Event data types
Flipped them
